--- src/test/java/com/vaadin/ui/checkbox/CheckBoxPageIT.java
@@ -1,41 +1,0 @@
-/*
- * Copyright 2000-2017 Vaadin Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.vaadin.ui.checkbox;
-
-import org.junit.Assert;
-import org.junit.Test;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebElement;
-
-import com.vaadin.flow.testutil.AbstractComponentIT;
-import com.vaadin.flow.testutil.TestPath;
-
-/**
- * @author Vaadin Ltd
- *
- */
-@TestPath("checkbox-test")
-public class CheckBoxPageIT extends AbstractComponentIT {
-
-    @Test
-    public void valueChangeListenerInSameRequest_checkBoxValueIsSet() {
-        open();
-
-        WebElement checkbox = findElement(By.id("checkbox"));
-        Assert.assertTrue(getInShadowRoot(checkbox, By.id("nativeCheckbox"))
-                .isSelected());
-    }
-}
